requirements.txt
pandas
alpaca
python-dotenv
boto3
pyarrow
pytorch-forecasting